Add files via upload
This script is used within the Bachelor Thesis report of Kasanmoesdiran (2025), a quote taken from the thesis describes the Python script and its functions.
 
"A replicatable unitcell is ran through a Python script which does the following:
• Finds all bonds using a cutoff distance method and stores them as a list of pairs of indices.
• Finds all molecules using the Depth First Search (DFS) algorithm with the pair indices as
a graph input. It then finds each unique molecule among the list of molecules by sorting
them by atom type.
This script takes in the replicatable unit cell .xyz file and the unit cell box information as
inputs. The unit cell box information can be obtained by following these steps:
• Load the replicatable unit cell’s .cif file obtained using Pymatgen into Mercury then
navigate to Display → More information → structure information, note down the
cell lengths and cell angles. These are the inputs for the molecule finder.

The script contains a function that can replicate any given unit cell, so as to avoid using ASE
for replication. This does, however, increase the amount of atoms and bonds that need to be
analyzed significantly, as the script currently contains O(N 2) complexity. This is typically not
necessary as LAMMPS has a built-in replicate function as well, but it can be useful if you are en-
countering issues with LAMMPS or simply want to see a list of all bonds in a replicated structure."